Scheduler: move testcafe status tests to jest
