+++ packages/devextreme/js/__internal/scheduler/__tests__/__mock__/model/toolbar.ts
@@ -1,0 +1,15 @@
+export class ToolbarModel {
+  element: HTMLDivElement | null;
+
+  constructor(element: HTMLDivElement | null) {
+    this.element = element;
+  }
+
+  getPrevButton(): HTMLDivElement | null | undefined {
+    return this.element?.querySelector('.dx-scheduler-navigator-previous');
+  }
+
+  getNextButton(): HTMLDivElement | null | undefined {
+    return this.element?.querySelector('.dx-scheduler-navigator-next');
+  }
+}
